Protect shutil.copytree call

+++ nsist/copymodules.py
@@ -162,7 +162,7 @@ def copy_distribution(modname, target):
         return
 
     egg_info_path = distribution._provider.egg_info
-    if os.path.exists(egg_info_path):
+    if os.path.exists(egg_info_path) and not os.path.exists(dest):
         dest = os.path.join(target, os.path.basename(egg_info_path))
         shutil.copytree(egg_info_path, dest)
